+++ react/ProductSearchContextProvider.js
@@ -1,28 +1,35 @@
-import PropTypes from 'prop-types'
 import React, { Component } from 'react'
+import { path } from 'ramda'
 
-import ProductSearchDataLayer from './components/ProductSearchDataLayer'
+import DataLayerApolloWrapper from './components/DataLayerApolloWrapper'
 import SearchQueryContainer from './components/SearchQueryContainer'
-import { searchQueryPropTypes } from './constants/propTypes'
+import { searchContextPropTypes } from './constants/propTypes'
 import { SearchQueryContext } from './constants/searchContext'
 
 class ProductSearchContextProvider extends Component {
-  static propTypes = {
-    params: PropTypes.shape({
-      /** Brand name */
-      brand: PropTypes.string,
+  static propTypes = searchContextPropTypes
 
-      /** handles /:department/d
-       *  or /:department/:category
-       *  or /:department/:category/:subcategory */
-      department: PropTypes.string,
-      category: PropTypes.string,
-      subcategory: PropTypes.string,
+  getData = () => {
+    const { searchQuery } = this.props
 
-      /** Search's term, e.g: eletronics. */
-      term: PropTypes.string,
-    }),
-    ...searchQueryPropTypes,
+    if (!searchQuery) {
+      return null
+    }
+
+    const { products } = searchQuery
+
+    return {
+      ecommerce: {
+        impressions: products.map((product, index) => ({
+          id: product.productId,
+          name: product.productName,
+          list: 'Search Results',
+          brand: product.brand,
+          category: path(['categories', '0'], product),
+          position: index + 1,
+        })),
+      },
+    }
   }
 
   render() {
@@ -41,13 +48,14 @@ class ProductSearchContextProvider extends Component {
       <SearchQueryContainer {...props}>
         <SearchQueryContext.Consumer>
           {contextProps => (
-            <ProductSearchDataLayer
-              searchQuery={contextProps.searchQuery}
+            <DataLayerApolloWrapper
+              getData={this.getData}
               loading={
                 contextProps.state.loading || contextProps.searchQuery.loading
-              }>
+              }
+            >
               {React.cloneElement(this.props.children, contextProps)}
-            </ProductSearchDataLayer>
+            </DataLayerApolloWrapper>
           )}
         </SearchQueryContext.Consumer>
       </SearchQueryContainer>