+++ CHANGELOG.md
@@ -9,6 +9,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Component `OrderFormContext` 
 
+## [1.3.2] - 2018-07-25
+### Changed
+- Delete all unnecessary path manipulation to not need pages information.
+
+## [1.3.1] - 2018-07-24
+### Fixed
+- Product search context throwing error on undefined prop.
+
+## [1.3.0] - 2018-07-24
+### Changed
+- `searchQuery` to retrieve facets and recordsFiltered.
+
+### Removed
+- `facetsQuery`.
+
+## [1.2.2] - 2018-07-24
+### Fixed
+- Data layer component props.
+
 ## [1.2.1] - 2018-07-23
 ### Added
 - Stale cache for product